ts content csp checks
