+++ src/js/content/updateCurrentState.ts
@@ -1,0 +1,18 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import {MESSAGE_TYPE, State} from '../config';
+
+export const currentOrigin = {val: ''};
+
+export function updateCurrentState(state: State) {
+  chrome.runtime.sendMessage({
+    type: MESSAGE_TYPE.UPDATE_STATE,
+    state,
+    origin: currentOrigin.val,
+  });
+}
